fix(discord/vuln): Changed config file to be split into server and client
Fixed a vuln related to how the config was shared to both the server and the client. The config has been split into two files now server.config.lua and client.config.lua, all sensitive information that is added for server side use should only be using the server.config.lua and the global ServerConfig table.